Introduce "locate current feature" & refactoring
Introduces a new button in the feature form that zooms and pans to the current
item. With a long press it will follow the current item.

This further introduces a new central access point for map settings available
to QML. This is still based on QgsMapCanvas as a bridge to modify the
QgsMapSettings but it is planned to make this the central hub and controller of
all map settings modifications.