keyHandler in mainWindow - removed onBackPressed in java-activity
the onBackPressed slot in the java-activity was a try to handle the back-key when it's pressed on the very start. because it didn't was handled at all. it still wasn't with the onBackPressed slot - so this is removed back again.

the keyHandler is for the same occation. Should handle the back-key hopefully and closes the app (with the two step closing).