focusstack
focusstack is a qlist that handles the items added to the stack (focusstack.addFocusTaker( this ))
if an item get's out of focus but still needs the focus (visible) it should receive it back when the next item get's out of focus on purpose.

additionally it does on backkey at featureform not close the featurelist bar - it goes back to the list instead.

still open issue is that you cannot select the featureformbar by clicking on it. This should be solved with the mouse area in FeatureListForm but there is the problem that you cannot select a widget anymore. so it's commented out...