Open PDF and other files on External Resource Widget (#860)
Displays values of the widget external resource that are not from mime type image/* as blue links-like labels.

Opens the files with QFieldOpenExternallyActivity and returns an error status. E.g. If no app is installed for a specific type it would throw a runtime error, this is cached.

Images and files are clickable in read only mode as well.

Geo information Badge is only visible when there is a file.

Image name of ic_photo_notavailable_black_24dp is set to correct color and size.