Track Lines and Polygons reloaded  (#855)
Tracking functionality started and stopped by pressing the layer in the legend

The tracking sessions are handled in the trackingmodel visualized over the Tracking in QML containing rubberband, tracker, layer and feature.

The tracker contains the technical tracking part with the parameters for the minimum distance and the (minimum) time between the tracked vertexes.

The rubberband contains functionalities to store the measure value. In the tracking case this is the time between the start of the tracking session and the current position time.

The layertreemapcanvasbridge handles the communication between layertreemodel and trackingmodel regarding if the layer is currently on a tracking session and the visibility of the layer.

The FeatureForm contains the parameter dontSave, since on tracking the attributes of the feature need to be set but not stored before the tracking started (since there is not yet a geometry).