Use expression configured in QFieldSync for picturenames (#790)
Uses expressions for image path.
Reads the customProperties for photo_naming from the project file
Uses an ExpressionEvaluator object and passes layer and project to use it with the scopes

Evaluate suffix of given filename and handle the image with this suffix. This means if the camera uses png and png is configured as a file suffix with expressions, it uses png. But when no expression is configured still jpg is used.