Rename save() function in FeatureForm.qml to confirm, because it's called when we confirm the add or edit of the feature by pressing a button, tipping on canvas and no saveing in the background (like auto save or buffer) - this function calls the save() function
The save() save function replaces the buffer() function. This is called when the feature has to be stored (in case of adding childs) and will maybe be deleted later again (in case the user cancels the add feature session)

The save() function is called for the auto save on value changed.

We don't have a bool called buffered in FeatureForm anymore. It checks if the feature has been stored already with FeatureModel::featureExists()