Enable closing of drawer by clicking on map or back key if auto save is active even if the constraints are not fulfilled. The features remain stored in the last valid state.
On canceling "add" feature on autosave, do not delete the feature again (it would only be deleted, when the feature has been saved for referencing childs and canceled again).

Channged toast message to "Last changes discarded", because it's more general and handles the case where some changes are stored (on auto save).