Performance boost when downloading packages
Use `md5` checksum instead of `sha256` checksum to gain extra performance
when downloading packaged files in QField. The `sha256` causes N+1 requests
to the S3 storage, therefore makes the QFieldCloud response on the 
`packages/<project-id>/latest` endpoint super slow.

This change will cause an extra initial download of the project file for
all users when they sync, since we renamed `lastProjectFileSha256` to
`lastProjectFileMd5`. IMO it is acceptable, otherwise we have to handle
extra code that I would prefer not to maintain.