Always track file-based layers in QFieldCloud
If the layer is configured as "NO_ACTION" for QFieldCloud, then this
layer is never tracked for changes in QFieldCloud. This means any 
changes on this layer will never reach QFieldCloud.

This development ignores whatever is configured as cloud action. When
the layer is file-based and it is not configured as "REMOVE", then the
layer will be tracked by QFielCloud.