Update CMakeLists.txt
I was trying to compile QField for Android using the Docker image. The whole process of dependencies built went well. Then, when configuring the project, CMake failed. It was related to the QT_HOST_PATH_CMAKE_DIR variable not being found by QtPublicDependencyHelpers.cmake.

By commenting the line where we manually set the variable, the process successfully finished.

Could it be that some change in Qt 6.7 fixed the previous bug of not finding the path? Now that it's fixed, CMake finds it automatically, so we shouldn't need to set the parameter manually…