Sanitize attachment filenames before saving so they are acceptable by QFieldCloud
QFieldCloud recently added a restriction for file names so they can work on all
popular operating systems. However, the attachment naming expressions have no
restrictions to what they are going to evaluate.

Therefore the new `FileUtils::sanitizeFilePath` function has been added
that will convert the disallowed chars into underscores and return an empty
string when the file path cannot be fixed, for example when it is too long.

If an empty string is returned,then the default naming kicks in to guarantee
the filename can be saved and can be uploaded to QFieldCloud.