fix(qfc): potential problem where a deleted delta scrambles the delta ids
Prior to this commit the data was stored as:

    // { localLayerId: { localPk: indexInTheDeltafile } }
    QMap<QString, QMap<QString, int>> mLocalPkDeltaIdx;

However, in `DeltaFileWrapper::mergeDeleteDelta` we have `mDeltas.removeAt`,
which can potentially change all delta indices in the deltafile, but never
update them in `mLocalPkDeltaIdx`.

This change removes the dependency of the positions in the deltafile and
adds replaces it with the uuids of the deltas:

    // { localLayerId: { localPk: deltaUuid } }
    QMap<QString, QMap<QString, QString>> mLocalPkToDeltaUuid;

As the UUIDs does not have a direct mapping, a new method is also added:

    /**
     * Retuns the index position of a delta with given \a uuid in the deltas list or -1 if missing.
     *
     * @param uuid the uuid we are looking for
     */
    int getDeltaIndexByUuid( const QString &uuid ) const;

Since the function must always return a value >=0, we added an assert
and no other special handling.